+++ packages/browser-wallet/src/popup/pages/Account/i18n/en.ts
@@ -325,10 +325,11 @@ const t = {
     transactionMessage: {
         configureBaker: {
             registerBaker:
-                'You are about to submit a register baker transaction that locks some of your funds in a stake. If you want to unlock the stake again, there will be a cool-down period.',
+                'You are about to submit a register baker transaction that locks some of your funds in a stake. If you want to unlock the stake again, there will be a cool-down period of {{ cooldownPeriod }} days.',
             lowerBakerStake:
-                'You are about to submit a transaction that lowers your baker stake. Lowering your stake has a cool-down period, meaning it will not take effect immediately.\n\nThe baker cannot be removed and the stake cannot be changed until the cool-down is over.',
-            removeBaker: 'Are you sure you want to make the following transaction to stop baking?',
+                'You are about to submit a baker transaction that lowers your  baker stake. It will take effect after {{ cooldownPeriod }} days and the baker stake cannot be changed during this period of time.',
+            removeBaker:
+                'Are you sure you want to make the following transaction to stop baking? It will take effect after {{ cooldownPeriod }} days and the baker stake cannot be changed during this period of time.',
             notice: {
                 start: 'Once your transaction has successfully finalized, the baker registration will take effect from the next pay day.\n\nRemember to restart your node with the baker keys.',
                 update: 'Once your transaction has successfully finalized, the baker update will take effect from the next pay day.',
@@ -338,10 +339,10 @@ const t = {
         },
         configureDelegation: {
             register:
-                'This will lock your delegation amount for at least {{ cooldownPeriod}} days from the time you remove or decrease your delegation.',
+                'You are about to submit a register delegation transaction that locks some of your funds in a stake. If you want to unlock the stake again, there will be a cool-down period of {{ cooldownPeriod }} days.',
             lowerDelegationStake:
                 'You are about to submit a delegation transaction that lowers your delegation amount. It will take effect after {{ cooldownPeriod }} days and the delegation amount cannot be changed during this period of time.',
-            remove: 'Are you sure you want to remove your delegation?',
+            remove: 'Are you sure you want to make the following transaction to stop delegation? It will take effect after {{ cooldownPeriod }} days and the delegation amount cannot be changed during this period of time.',
             notice: {
                 start: 'Once your transaction has successfully finalized, the delegation will take effect from the next pay day.',
                 update: 'Once your transaction has successfully finalized, the delegation update will take effect from the next pay day.',